@@ -2,12 +2,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<artifactId>spring-cloud-stream-binder-kafka-parent</artifactId>
-	<version>3.2.2-SNAPSHOT</version>
+	<version>4.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
 		<artifactId>spring-cloud-build</artifactId>
-		<version>3.1.0</version>
+		<version>4.0.0-SNAPSHOT</version>
 		<relativePath />
 	</parent>
 	<scm>
@@ -20,11 +20,11 @@
 		<tag>HEAD</tag>
 	</scm>
 	<properties>
-		<java.version>1.8</java.version>
-		<spring-kafka.version>2.8.0</spring-kafka.version>
-		<spring-integration-kafka.version>5.5.5</spring-integration-kafka.version>
+		<java.version>17</java.version>
+		<spring-kafka.version>3.0.0-SNAPSHOT</spring-kafka.version>
+		<spring-integration-kafka.version>6.0.0-SNAPSHOT</spring-integration-kafka.version>
 		<kafka.version>3.0.0</kafka.version>
-		<spring-cloud-stream.version>3.2.2-SNAPSHOT</spring-cloud-stream.version>
+		<spring-cloud-stream.version>4.0.0-SNAPSHOT</spring-cloud-stream.version>
 		<maven-checkstyle-plugin.failsOnError>true</maven-checkstyle-plugin.failsOnError>
 		<maven-checkstyle-plugin.failsOnViolation>true</maven-checkstyle-plugin.failsOnViolation>
 		<maven-checkstyle-plugin.includeTestSourceDirectory>true</maven-checkstyle-plugin.includeTestSourceDirectory>
@@ -144,7 +144,7 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-antrun-plugin</artifactId>
-					<version>1.7</version>
+<!--					<version>1.7</version>-->
 				</plugin>
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
