@@ -18,6 +18,8 @@ package org.springframework.cloud.stream.binder.kafka.streams.integration;
 
 import java.time.Duration;
 import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.Consumer;
 
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.kstream.KStream;
@@ -32,11 +34,6 @@ import org.junit.Test;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.cloud.stream.annotation.StreamListener;
-import org.springframework.cloud.stream.binder.kafka.streams.annotations.KafkaStreamsStateStore;
-import org.springframework.cloud.stream.binder.kafka.streams.properties.KafkaStreamsStateStoreProperties;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
@@ -67,6 +64,7 @@ public class KafkaStreamsStateStoreIntegrationTests {
 		app.setWebApplicationType(WebApplicationType.NONE);
 		ConfigurableApplicationContext context = app.run("--server.port=0",
 				"--spring.jmx.enabled=false",
+				"--spring.cloud.stream.function.bindings.process-in-0=input",
 				"--spring.cloud.stream.bindings.input.destination=foobar",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.commit.interval.ms=1000",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.default.key.serde"
@@ -90,40 +88,13 @@ public class KafkaStreamsStateStoreIntegrationTests {
 	}
 
 	@Test
-	public void testKstreamStateStoreBuilderBeansDefinedInApplication() throws Exception {
-		SpringApplication app = new SpringApplication(StateStoreBeanApplication.class);
-		app.setWebApplicationType(WebApplicationType.NONE);
-		ConfigurableApplicationContext context = app.run("--server.port=0",
-				"--spring.jmx.enabled=false",
-				"--spring.cloud.stream.bindings.input3.destination=foobar",
-				"--spring.cloud.stream.kafka.streams.binder.configuration.commit.interval.ms=1000",
-				"--spring.cloud.stream.kafka.streams.binder.configuration.default.key.serde"
-						+ "=org.apache.kafka.common.serialization.Serdes$StringSerde",
-				"--spring.cloud.stream.kafka.streams.binder.configuration.default.value.serde"
-						+ "=org.apache.kafka.common.serialization.Serdes$StringSerde",
-				"--spring.cloud.stream.kafka.streams.bindings.input3.consumer.applicationId"
-						+ "=KafkaStreamsStateStoreIntegrationTests-xyzabc-123",
-				"--spring.cloud.stream.kafka.streams.binder.brokers="
-						+ embeddedKafka.getBrokersAsString());
-		try {
-			Thread.sleep(2000);
-			receiveAndValidateFoo(context, StateStoreBeanApplication.class);
-		}
-		catch (Exception e) {
-			throw e;
-		}
-		finally {
-			context.close();
-		}
-	}
-
-
-	@Test
 	public void testSameStateStoreIsCreatedOnlyOnceWhenMultipleInputBindingsArePresent() throws Exception {
 		SpringApplication app = new SpringApplication(ProductCountApplicationWithMultipleInputBindings.class);
 		app.setWebApplicationType(WebApplicationType.NONE);
 		ConfigurableApplicationContext context = app.run("--server.port=0",
 				"--spring.jmx.enabled=false",
+				"--spring.cloud.stream.function.bindings.process-in-0=input1",
+				"--spring.cloud.stream.function.bindings.process-in-1=input2",
 				"--spring.cloud.stream.bindings.input1.destination=foobar",
 				"--spring.cloud.stream.bindings.input2.destination=hello-foobar",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.commit.interval.ms=1000",
@@ -171,22 +142,12 @@ public class KafkaStreamsStateStoreIntegrationTests {
 			assertThat(state.persistent()).isTrue();
 			assertThat(productCount.processed).isTrue();
 		}
-		else if (clazz.isAssignableFrom(StateStoreBeanApplication.class)) {
-			StateStoreBeanApplication productCount = context
-					.getBean(StateStoreBeanApplication.class);
-			WindowStore<Object, String> state = productCount.state;
-			assertThat(state != null).isTrue();
-			assertThat(state.name()).isEqualTo("mystate");
-			assertThat(state.persistent()).isTrue();
-			assertThat(productCount.processed).isTrue();
-		}
 		else {
-			fail("Expected assertiond did not happen");
+			fail("Expected assertions did not happen");
 		}
 
 	}
 
-	@EnableBinding(KafkaStreamsProcessorX.class)
 	@EnableAutoConfiguration
 	public static class ProductCountApplication {
 
@@ -194,46 +155,10 @@ public class KafkaStreamsStateStoreIntegrationTests {
 
 		boolean processed;
 
-		@StreamListener("input")
-		@KafkaStreamsStateStore(name = "mystate", type = KafkaStreamsStateStoreProperties.StoreType.WINDOW, lengthMs = 300000, retentionMs = 300000)
-		@SuppressWarnings({ "deprecation", "unchecked" })
-		public void process(KStream<Object, Product> input) {
+		@Bean
+		public Consumer<KStream<Object, Product>> process() {
 
-			input.process(() -> new Processor<Object, Product>() {
-
-				@Override
-				public void init(ProcessorContext processorContext) {
-					state = (WindowStore) processorContext.getStateStore("mystate");
-				}
-
-				@Override
-				public void process(Object s, Product product) {
-					processed = true;
-				}
-
-				@Override
-				public void close() {
-					if (state != null) {
-						state.close();
-					}
-				}
-			}, "mystate");
-		}
-	}
-
-	@EnableBinding(KafkaStreamsProcessorZ.class)
-	@EnableAutoConfiguration
-	public static class StateStoreBeanApplication {
-
-		WindowStore<Object, String> state;
-
-		boolean processed;
-
-		@StreamListener("input3")
-		@SuppressWarnings({"unchecked" })
-		public void process(KStream<Object, Product> input) {
-
-			input.process(() -> new Processor<Object, Product>() {
+			return input -> input.process(() -> new Processor<Object, Product>() {
 
 				@Override
 				public void init(ProcessorContext processorContext) {
@@ -263,8 +188,6 @@ public class KafkaStreamsStateStoreIntegrationTests {
 		}
 	}
 
-
-	@EnableBinding(KafkaStreamsProcessorY.class)
 	@EnableAutoConfiguration
 	public static class ProductCountApplicationWithMultipleInputBindings {
 
@@ -272,33 +195,41 @@ public class KafkaStreamsStateStoreIntegrationTests {
 
 		boolean processed;
 
-		@StreamListener
-		@KafkaStreamsStateStore(name = "mystate", type = KafkaStreamsStateStoreProperties.StoreType.WINDOW, lengthMs = 300000, retentionMs = 300000)
-		@SuppressWarnings({ "deprecation", "unchecked" })
-		public void process(@Input("input1")KStream<Object, Product> input, @Input("input2")KStream<Object, Product> input2) {
+		@Bean
+		public BiConsumer<KStream<Object, Product>, KStream<Object, Product>> process() {
 
-			input.process(() -> new Processor<Object, Product>() {
+			return (input, input2) -> {
 
-				@Override
-				public void init(ProcessorContext processorContext) {
-					state = (WindowStore) processorContext.getStateStore("mystate");
-				}
+				input.process(() -> new Processor<Object, Product>() {
 
-				@Override
-				public void process(Object s, Product product) {
-					processed = true;
-				}
-
-				@Override
-				public void close() {
-					if (state != null) {
-						state.close();
+					@Override
+					public void init(ProcessorContext processorContext) {
+						state = (WindowStore) processorContext.getStateStore("mystate");
 					}
-				}
-			}, "mystate");
 
-			//simple use of input2, we are not using input2 for anything other than triggering some test behavior.
-			input2.foreach((key, value) -> { });
+					@Override
+					public void process(Object s, Product product) {
+						processed = true;
+					}
+
+					@Override
+					public void close() {
+						if (state != null) {
+							state.close();
+						}
+					}
+				}, "mystate");
+				//simple use of input2, we are not using input2 for anything other than triggering some test behavior.
+				input2.foreach((key, value) -> { });
+			};
+		}
+
+		@Bean
+		public StoreBuilder mystore() {
+			return Stores.windowStoreBuilder(
+					Stores.persistentWindowStore("mystate",
+							Duration.ofMillis(3), Duration.ofMillis(3), false), Serdes.String(),
+					Serdes.String());
 		}
 	}
 
@@ -314,26 +245,5 @@ public class KafkaStreamsStateStoreIntegrationTests {
 			this.id = id;
 		}
 
-	}
-
-	interface KafkaStreamsProcessorX {
-
-		@Input("input")
-		KStream<?, ?> input();
-	}
-
-	interface KafkaStreamsProcessorY {
-
-		@Input("input1")
-		KStream<?, ?> input1();
-
-		@Input("input2")
-		KStream<?, ?> input2();
-	}
-
-	interface KafkaStreamsProcessorZ {
-
-		@Input("input3")
-		KStream<?, ?> input3();
 	}
 }
