@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 the original author or authors.
+ * Copyright 2018-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
 
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.common.TopicPartition;
@@ -33,10 +34,6 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.cloud.stream.annotation.Output;
-import org.springframework.cloud.stream.annotation.StreamListener;
 import org.springframework.cloud.stream.binder.Binder;
 import org.springframework.cloud.stream.binder.BinderFactory;
 import org.springframework.cloud.stream.binder.ConsumerProperties;
@@ -47,10 +44,9 @@ import org.springframework.cloud.stream.binder.kafka.properties.KafkaConsumerPro
 import org.springframework.cloud.stream.binder.kafka.properties.KafkaProducerProperties;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.test.rule.EmbeddedKafkaRule;
 import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.SubscribableChannel;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
 
@@ -62,6 +58,11 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE, properties = {
+		"spring.cloud.stream.function.definition=process;processCustom",
+		"spring.cloud.stream.function.bindings.process-in-0=standard-in",
+		"spring.cloud.stream.function.bindings.process-out-0=standard-out",
+		"spring.cloud.stream.function.bindings.processCustom-in-0=custom-in",
+		"spring.cloud.stream.function.bindings.processCustom-out-0=custom-out",
 		"spring.cloud.stream.kafka.bindings.standard-out.producer.configuration.key.serializer=FooSerializer.class",
 		"spring.cloud.stream.kafka.default.producer.configuration.key.serializer=BarSerializer.class",
 		"spring.cloud.stream.kafka.default.producer.configuration.value.serializer=BarSerializer.class",
@@ -167,42 +168,23 @@ public class KafkaBinderExtendedPropertiesTest {
 				Boolean.TRUE);
 	}
 
-	@EnableBinding(CustomBindingForExtendedPropertyTesting.class)
 	@EnableAutoConfiguration
+	@Configuration
 	public static class KafkaMetricsTestConfig {
 
-		@StreamListener("standard-in")
-		@SendTo("standard-out")
-		public String process(String payload) {
-			return payload;
+		@Bean
+		public Function<String, String> process() {
+			return payload -> payload;
 		}
 
-		@StreamListener("custom-in")
-		@SendTo("custom-out")
-		public String processCustom(String payload) {
-			return payload;
-		}
-
+	@Bean
+	public Function<String, String> processCustom() {
+		return payload -> payload;
+	}
 		@Bean
 		public RebalanceListener rebalanceListener() {
 			return new RebalanceListener();
 		}
-
-	}
-
-	interface CustomBindingForExtendedPropertyTesting {
-
-		@Input("standard-in")
-		SubscribableChannel standardIn();
-
-		@Output("standard-out")
-		MessageChannel standardOut();
-
-		@Input("custom-in")
-		SubscribableChannel customIn();
-
-		@Output("custom-out")
-		MessageChannel customOut();
 
 	}
 
@@ -215,23 +197,18 @@ public class KafkaBinderExtendedPropertiesTest {
 		@Override
 		public void onPartitionsRevokedBeforeCommit(String bindingName,
 				Consumer<?, ?> consumer, Collection<TopicPartition> partitions) {
-
 		}
 
 		@Override
 		public void onPartitionsRevokedAfterCommit(String bindingName,
 				Consumer<?, ?> consumer, Collection<TopicPartition> partitions) {
-
 		}
 
 		@Override
 		public void onPartitionsAssigned(String bindingName, Consumer<?, ?> consumer,
 				Collection<TopicPartition> partitions, boolean initial) {
-
 			this.bindings.put(bindingName, initial);
 			this.latch.countDown();
 		}
-
 	}
-
 }
