@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 the original author or authors.
+ * Copyright 2018-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,8 @@
 
 package org.springframework.cloud.stream.binder.kafka.integration.topic.configs;
 
+import java.util.function.Function;
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -23,24 +25,21 @@ import org.junit.runner.RunWith;
 
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.cloud.stream.annotation.Output;
-import org.springframework.cloud.stream.annotation.StreamListener;
+import org.springframework.context.annotation.Bean;
 import org.springframework.kafka.test.rule.EmbeddedKafkaRule;
-import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.SubscribableChannel;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
 
 /**
  * @author Heiko Does
+ * @author Soby Chacko
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest(
 		classes = BaseKafkaBinderTopicPropertiesUpdateTest.TopicAutoConfigsTestConfig.class,
 		webEnvironment = SpringBootTest.WebEnvironment.NONE, properties = {
+		"spring.cloud.stream.function.bindings.process-in-0=standard-in",
+		"spring.cloud.stream.function.bindings.process-out-0=standard-out",
 		"spring.cloud.stream.kafka.bindings.standard-out.producer.topic.properties.retention.ms=9001",
 		"spring.cloud.stream.kafka.default.producer.topic.properties.retention.ms=-1",
 		"spring.cloud.stream.kafka.bindings.standard-in.consumer.topic.properties.retention.ms=9001",
@@ -65,24 +64,12 @@ public abstract class BaseKafkaBinderTopicPropertiesUpdateTest {
 		System.clearProperty(KAFKA_BROKERS_PROPERTY);
 	}
 
-	@EnableBinding(CustomBindingForTopicPropertiesUpdateTesting.class)
 	@EnableAutoConfiguration
 	public static class TopicAutoConfigsTestConfig {
 
-		@StreamListener("standard-in")
-		@SendTo("standard-out")
-		public String process(String payload) {
-			return payload;
+		@Bean
+		public Function<String, String> process() {
+			return payload -> payload;
 		}
 	}
-
-	interface CustomBindingForTopicPropertiesUpdateTesting {
-
-		@Input("standard-in")
-		SubscribableChannel standardIn();
-
-		@Output("standard-out")
-		MessageChannel standardOut();
-	}
-
 }
